chore: rename package storage to database
