+++ warehouse-management-service/internal/server/server.go
@@ -1,25 +1,25 @@
 package server
 
 import (
-	"warehouse-management-service/pkg/storage"
+	"warehouse-management-service/pkg/database"
 
 	"github.com/go-chi/chi/middleware"
 	"github.com/go-chi/chi/v5"
 )
 
 type Server struct {
-	storage storage.Service
-	Router  *chi.Mux
+	db     database.Service
+	Router *chi.Mux
 }
 
 // Creates and returns a chi Router and configures the server routes
-func New(storage storage.Service) *Server {
+func New(db database.Service) *Server {
 	router := chi.NewRouter()
 	router.Use(middleware.Logger)
 
 	server := &Server{
-		Router:  router,
-		storage: storage,
+		Router: router,
+		db:     db,
 	}
 	server.routes()
 	return server
