+++ warehouse-management-service/pkg/database/database.go
@@ -1,4 +1,4 @@
-package storage
+package database
 
 type Service interface {
 	Close() error
