+++ warehouse-management-service/pkg/database/postgres/postgres.go
@@ -3,7 +3,7 @@ package postgres
 import (
 	"database/sql"
 	"fmt"
-	storage "warehouse-management-service/pkg/storage"
+	"warehouse-management-service/pkg/database"
 
 	// Loads postgres drivers
 	_ "github.com/lib/pq"
@@ -13,7 +13,7 @@ type postgres struct {
 	db *sql.DB
 }
 
-func New(host, port, user, password, dbName string, sslMode bool) (storage.Service, error) {
+func New(host, port, user, password, dbName string, sslMode bool) (database.Service, error) {
 	connect := fmt.Sprintf("host=%s port=%s user=%s password=%s dbname=%s sslmode=disable",
 		host, port, user, password, dbName)
 	db, err := sql.Open("postgres", connect)
